chore: factor out logic for determining source files from webpack config
